Issue#162: Reducing the size of the item factory by outsourcing the methods in seperated files. Also rearranging the imports.
