+++ tests/test_item_brand.py
@@ -1,5 +1,6 @@
 """Class to test the item asin."""
 import unittest
+
 from crawler.item_factory import item_factory
