+++ tests/test_item_factory.py
@@ -1,5 +1,6 @@
 """Class to test the item factory module."""
 import unittest
+
 from crawler.item_factory import item_factory
