+++ tests/test_item_seller.py
@@ -1,5 +1,6 @@
 """Class to test the item seller."""
 import unittest
+
 from crawler.item_factory import item_factory
