+++ tests/test_item_url.py
@@ -1,5 +1,6 @@
 """Class to test the item url."""
 import unittest
+
 from crawler.item_factory import item_factory
